- Code cleanup and added junit tests
