+++ modules/apis/src/main/java/com/axway/apim/api/export/model/APICert.java
@@ -2,14 +2,14 @@ package com.axway.apim.api.export.model;
 
 public class APICert {
 
-    private String id;
-    private String name;
-    private String path;
-    private String version;
-    private String commonName;
-    private long validAfter;
-    private long validBefore;
-    private String md5FingerPrint;
+    private final String id;
+    private final String name;
+    private final String path;
+    private final String version;
+    private final String commonName;
+    private final long validAfter;
+    private final long validBefore;
+    private final String md5FingerPrint;
 
     public APICert(String id, String name, String path, String version, String commonName, long validAfter, long validBefore, String md5FingerPrint) {
         this.id = id;
@@ -26,63 +26,37 @@ public class APICert {
         return id;
     }
 
-    public void setId(String id) {
-        this.id = id;
-    }
-
     public String getName() {
         return name;
     }
 
-    public void setName(String name) {
-        this.name = name;
-    }
 
     public String getPath() {
         return path;
     }
 
-    public void setPath(String path) {
-        this.path = path;
-    }
 
     public String getVersion() {
         return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
     }
 
     public String getCommonName() {
         return commonName;
     }
 
-    public void setCommonName(String commonName) {
-        this.commonName = commonName;
-    }
 
     public long getValidAfter() {
         return validAfter;
     }
 
-    public void setValidAfter(long validAfter) {
-        this.validAfter = validAfter;
-    }
 
     public long getValidBefore() {
         return validBefore;
     }
 
-    public void setValidBefore(long validBefore) {
-        this.validBefore = validBefore;
-    }
 
     public String getMd5FingerPrint() {
         return md5FingerPrint;
     }
 
-    public void setMd5FingerPrint(String md5FingerPrint) {
-        this.md5FingerPrint = md5FingerPrint;
-    }
 }
